+++ Assets/SequenceSDK/SequenceIndexer/Scripts/DataTypes/Transaction.cs
@@ -1,12 +1,14 @@
+using System.Numerics;
+
 namespace SequenceSharp
 {
     [System.Serializable]
     public class Transaction
     {
         public string txnHash;
-        public int blockNumber;
+        public BigInteger blockNumber;
         public string blockHash;
-        public int chainId;
+        public BigInteger chainId;
         public string metaTxnID;
         public TxnTransfer[] transfers;
         public string timestamp;
